refactor(button): simplify styles and update dependencies
Refactored button styles for cleaner and more precise variant handling. Replaced and optimized imports for consistency across components. Updated package
