+++ package.json
@@ -59,6 +59,7 @@
     "@fontsource/roboto": "^5.1.1",
     "@fortawesome/fontawesome-svg-core": "^6.7.2",
     "@fortawesome/free-brands-svg-icons": "^6.7.2",
+    "@fortawesome/free-regular-svg-icons": "^6.7.2",
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@material/material-color-utilities": "^0.3.0",
     "@material/ripple": "^14.0.0",
